+++ test/libsolidity/syntaxTests/immutable/double_specifier.sol
@@ -3,5 +3,5 @@ contract C {
     uint immutable constant x;
 }
 // ----
-// ParserError: (32-41): Constantness already set to "immutable"
-// ParserError: (64-72): Constantness already set to "immutable"
+// ParserError: (32-41): Mutability already set to "immutable"
+// ParserError: (64-72): Mutability already set to "immutable"